Minor update
Google scholar link is now set to search the article title instead of searching doi in order to get better chance of find the article. Note that for articles with very long title, space might be lost occasionally due to the primary combination step. Not sure how to trouble shoot that but google is usually smart enough to suggest the right version.